chore(cfg): use long values for timeouts
related to zeebe-io/zeebe#815

+++ raft/src/main/java/io/zeebe/raft/state/Heartbeat.java
@@ -22,10 +22,10 @@ import io.zeebe.util.sched.clock.ActorClock;
 
 public class Heartbeat
 {
-    private final int electionInterval;
+    private final long electionInterval;
     private long lastHeartbeat = 0;
 
-    public Heartbeat(int electionInterval)
+    public Heartbeat(long electionInterval)
     {
         this.electionInterval = electionInterval;
     }
